# pkg/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CMDB API
  version: 0.1.0
  description: 纯自研 CMDB 核心 API 契约，契约先行，前后端据此并行开发
servers:
  - url: /
    description: 相对路径（与后端同域部署）
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/governance/rules:
    get:
      tags: [governance]
      summary: 查询稽核规则列表
      operationId: listAuditRules
      responses:
        '200':
          description: 稽核规则列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRuleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [governance]
      summary: 创建稽核规则
      operationId: createAuditRule
      description: 声明式规则 = 模型过滤条件（filter 属性等值）+ 断言表达式 + 待办模板；断言语法见 AuditRule.assertion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRuleRequest'
      responses:
        '200':
          description: 创建成功，返回规则
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/governance/rules/{rule_id}:
    patch:
      tags: [governance]
      summary: 修改稽核规则
      operationId: patchAuditRule
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRuleRequest'
      responses:
        '200':
          description: 修改成功，返回规则
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/governance/rules/{rule_id}/run:
    post:
      tags: [governance]
      summary: 手动执行稽核规则
      operationId: runAuditRule
      description: 立即执行一次规则：断言失败的 CI 按 (rule_id, ci_id) 去重生成待办；修复的 CI 待办自动关闭；dry_run 规则只出报告不落库
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 执行报告
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRuleRunResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/governance/todos:
    get:
      tags: [governance]
      summary: 查询整改待办列表
      operationId: listGovernanceTodos
      parameters:
        - name: status
          in: query
          description: 按状态过滤（缺省返回全部）
          required: false
          schema:
            type: string
            enum: [open, closed]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 整改待办分页列表（附规则名，最新在前）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceTodoListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/governance/todos/{todo_id}/close:
    post:
      tags: [governance]
      summary: 关闭整改待办
      operationId: closeGovernanceTodo
      description: 人工关闭待办（幂等）；规则下次执行通过时也会自动关闭
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 关闭成功，返回待办
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceTodo'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: meridian_token
      description: 登录成功后由服务端写入的 httpOnly JWT 会话 cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 采集器等非浏览器客户端使用的 Bearer JWT(登录接口响应体中的 token)
  parameters:
...
    Page:
      name: page
      in: query
      description: 页码，从 1 开始
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: 每页条数
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  responses:
    BadRequest:
      description: 请求参数或校验规则不通过
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未登录或会话已失效(错误码 UNAUTHORIZED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 已登录但缺少所需权限点(错误码 FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 机器可读错误码
          example: VALIDATION_FAILED
        message:
          type: string
          description: 人类可读错误信息
          example: 属性 hostname 为必填项
        details:
          type: object
          additionalProperties: true
          description: 逐字段校验错误等补充信息
...
    AuditRule:
      type: object
      description: |
        稽核规则（F-081）。assertion 为声明式断言表达式，语法：
        比较（==/!=/>/>=/</<=/contains）、逻辑（and/or/not/括号）、
        函数（empty(attr)/not_empty(attr)/age_days(attr)/unique(attr)/biz_attributed()）、
        裸属性名真值判断（非空即真）。filter 为 CI 属性等值过滤（如 {"env":"prod"}）。
      required: [id, name, model_code, filter, assertion, message, enabled, dry_run]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        model_code:
          type: string
          description: 目标模型编码（如 host）
        type:
          type: string
          enum: [audit, auto_ingest]
          default: audit
          description: |
            规则类型：audit 为稽核规则（默认）；auto_ingest 为自动入库白名单（F-034）——
            调和判定 create 且模型+属性命中（filter 等值 + assertion 为真）时直接建档
            status=active，不写发现池；dry_run=true 的白名单规则只出报告不参与放行。
        filter:
          type: object
          description: CI 属性等值过滤条件，空对象表示模型内全部 CI
        assertion:
          type: string
          example: not_empty(cluster_name) and backup_count > 0
        message:
          type: string
          description: 违规时生成待办的标题
        enabled:
          type: boolean
        dry_run:
          type: boolean
          description: 演练模式：只出报告不产生/关闭待办
        last_run_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuditRuleRequest:
      type: object
      description: 规则创建/修改请求（PATCH 时全字段可选）
      required: [name, model_code, assertion, message]
      properties:
        name:
          type: string
        model_code:
          type: string
        type:
          type: string
          enum: [audit, auto_ingest]
          description: 规则类型，缺省 audit；auto_ingest 为自动入库白名单（F-034）
        filter:
          type: object
        assertion:
          type: string
        message:
          type: string
        enabled:
          type: boolean
        dry_run:
          type: boolean
    AuditRuleListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditRule'
...
    AuditRuleViolation:
      type: object
      required: [ci_id, title]
      properties:
        ci_id:
          type: string
          format: uuid
        title:
          type: string
    AuditRuleRunResult:
      type: object
      required: [rule_id, rule_name, dry_run, checked, violations, todos_created, todos_closed]
      properties:
        rule_id:
          type: string
          format: uuid
        rule_name:
          type: string
        dry_run:
          type: boolean
        checked:
          type: integer
          description: 参与检查的 CI 数
        violations:
          type: array
          items:
            $ref: '#/components/schemas/AuditRuleViolation'
        todos_created:
          type: integer
        todos_closed:
          type: integer
    GovernanceTodo:
      type: object
      required: [id, rule_id, ci_id, title, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        rule_id:
          type: string
          format: uuid
        rule_name:
          type: string
          description: 列表接口附带的规则名
        ci_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [open, closed]
        created_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
    GovernanceTodoListResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GovernanceTodo'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...